Refactor: MOD extract connection
패키지분리
- jdbc: jdbc 관련코드
- test: jdbc 테스트코드

refactoring
- get connection 분리
